Changes made to version 2
Trial list is now defined based on the session info provided.
The frame is shuffled and ordered based on a custom order defined in orderDict.
This will allow for all stims and sections to be shuffled and arranged randomly in various runs.

There is a bug at line 357:
if __name__ == '__main__':