Changes made to version 1 and 2
Version 1 now has a fixation cross for the duration of prime and target.
Fixation remains for 2.7 seconds post target.
Audio is now played as soon as cross appears.

Version 2 now has no fixation at all for the prime.
Fixation appears for the target and audio is played after a 1.5 seconds pause and remains for 2.7 seconds post target.

Bug fixed - The breaks now appear every 25 trials or every three trials for the test rotation.
Fix is in the if block where a new condition was added to check the contKey list for return key press rather than use another if block.
The use of continue was causing the rest of the loop to be skipped.