Final rotations created
Version 1 and 2 now have all possible subgroups and versions added.
All stim_lists are now in the directory 'stim_lists' and paths have been updated.

The creation of the trial_list by reading in one the stim lists has been greatly simplified.
Stim lists are now read in with the function trial_list_reader, which creates trial_list based on the exp setup.

Minor modifications to the name and positions of the dialogue box.
An option to select either 'mouse' or 'eyelink' as the tracker has been added to the Experiment Setup dialogue.

Calibration points have been added after each break period to allow for recalibration of the tracker.