Minor changes to break format and instructions
After each break the participant is now asked to inform the experimenter to continue.
Press '1' to move on.
This should bring up the calibration window, which is not working as expected for eyelink in this version, but works for mouse gaze.

This is followed by the original 'Press enter to continue screen'.